Add files via upload
This is the solution to regulate the saline to grow whiteleg shrimp farms using the data of the Lerma River and the lakes of Chapala and Cuitzeo. If other variables are required then it is important to do the new calculations with the new data considering the salinity for the new species. Here are the conditions required for each species:
https://www.inapesca.gob.mx/portal/documentos/transparencia/transparenciafocalizada/cna-06062012/.